Signal 'schedule_changed' at the view layer
This fixes two issues:

* When the 'poll_now' signal was sent from Feed.save() it was inside the
  transaction. Thus the polling routine could (and usually would on my
  fast desktop machine) fail to see the updated next_check field value.
* Sending the signal from Feed.save() could cause unnecessary polling
  attempts as polling actually calls that method!

The signal was also renamed to 'schedule_changed' from 'poll_now' as it
is now emitted in more circumstances.
